显示股票的涨跌颜色，涨为红色，跌为绿色

+++ Plugins/Stock/Common.cpp
@@ -164,7 +164,7 @@ std::vector<std::string> CCommon::split(std::string str, std::string pattern)
     std::string::size_type pos;
     std::vector<std::string> result;
     str += pattern;//扩展字符串以方便操作
-    int size = str.size();
+    int size = static_cast<int>(str.size());
     for (int i = 0; i < size; i++)
     {
         pos = str.find(pattern, i);
